Discharge date correction
Using Nathaniel Fergurson from Onpoint suggestion to fix discharge dates. Pre-2000 values for the discharge date are likely default values that should not be present. I
For a temporary fix, recommended an approach similar to what Onpoint uses for constructing the inpatient_stay_summary table. There, we fall back to last_service_dt when discharge_dt is missing. I would also recommend this replacement when the discharge_dt is obviously invalid, such as when the discharge_dt is before the admission_dt. We typically see much more consistent reporting in the last_service_dt field.